+++ src/hooksFactory.js
@@ -1,19 +1,34 @@
 
 import { useSelector, useDispatch } from 'react-redux';
 
-export const includeHooks = ({
+export default (objectName, { byKey, parent }, {
   mapToProps,
   mapToPropsStripped,
   actions,
   actionsStripped,
 }) => {
-  
-  const getUseFactory = stripped => (props = {}) => {
-    const dispatch = useDispatch()
+  console.log(objectName, { actions, actionsStripped })
+  const getUseFactory = stripped => ({ [byKey]: id, [parent]: parentId, ...restProps } = {}) => {
+    if (!parent && typeof parentId !== 'undefined') {
+      console.error(
+        `Unrecognised parent prop ${
+          parent !== 'parent' ? `(${parent}) ` : ''
+        }while parent is not set given to ${objectName} hook.`
+      )
+    }
+    if (Object.values(restProps).length !== 0) {
+      console.error(
+        `Unrecognized props given to ${objectName} hook:`,
+        restProps
+      );
+    }
     const obj = useSelector(state => 
-      stripped ? mapToPropsStripped(state, props) : mapToProps(state, props)
+      stripped
+        ? mapToPropsStripped(state, { [byKey]: id, [parent]: parentId })
+        : mapToProps(state, { [byKey]: id, [parent]: parentId })
     );
 
+    const dispatch = useDispatch()
     return {
       ...obj,
       ...Object.fromEntries(